Refactor: move DateFormatter to util core module
